staticd: kill static_memory.h, use MTYPE_STATIC
This one needed a move of zebra_stable_node_cleanup() from static_vrf.c
to static_routes.c.  But it seems to actually make sense there.
