+++ staticd/static_routes.h
@@ -22,6 +22,9 @@
 
 #include "lib/mpls.h"
 #include "table.h"
+#include "memory.h"
+
+DECLARE_MGROUP(STATIC);
 
 /* Static route label information */
 struct static_nh_label {
@@ -197,6 +200,9 @@ extern bool static_add_nexthop_validate(const char *nh_vrf_name,
 					struct ipaddr *ipaddr);
 extern struct stable_info *static_get_stable_info(struct route_node *rn);
 extern void static_route_info_init(struct static_route_info *si);
+
+extern void zebra_stable_node_cleanup(struct route_table *table,
+				      struct route_node *node);
 
 /*
  * Max string return via API static_get_nh_str in size_t
